Norm: redirection_execute_part.c / pipe.c / pipe_utils.c / check_command.c / check_command_utils.c / print_free.c
